edit world - server -> client done

+++ WorldBuilder/WorldBuilder/Controllers/ColorPaletteController.cs
@@ -15,7 +15,7 @@ namespace WorldBuilder.Controllers
         }
 
         [HttpPost]
-        public async Task<ActionResult> _PostColor(int r, int g, int b, float a)
+        public async Task<ActionResult> _PostColor(float r, float g, float b, float a)
         {
             int res = await new ColorController().PostAsync(r, g, b, a);
             return Json(res);
@@ -24,7 +24,7 @@ namespace WorldBuilder.Controllers
         [HttpGet]
         public async Task<ActionResult> _GetColors(int offset, int ammount)
         {
-            List<Tuple<string, byte, byte, byte, float>> res = await new ColorController().GetAsync(offset, ammount);
+            List<Tuple<string, float, float, float, float>> res = await new ColorController().GetAsync(offset, ammount);
             return Json(res);
         }
